move all people to constants
